projectile/requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
colorama==0.4.6
Django==5.1.6
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-phonenumber-field==8.0.0
django-simple-history==3.8.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
phonenumbers==8.13.55
platformdirs==4.3.6
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.1
